Move validators and add flight number form validation
