+++ src/utils/validators.js
@@ -1,3 +1,7 @@
+export const isValidFlightDate = (value) => value instanceof Date
+
+export const isValidFlightNumber = (value) => /([A-Z]+)([0-9]+)/.test(value)
+
 export const isValidFlight = ({ arrival, date, departure, flightNumber, passengers, type }) => {
   if (type === 'locations') {
     return !!departure && !!arrival && passengers
